minor cleanup and fixups
